% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjm.R
\name{rjm}
\alias{rjm}
\title{Function rjm}
\usage{
rjm(
  y,
  X,
  K = 2,
  outer = 10,
  inner = 20,
  parallel = FALSE,
  epsilon = 1e-06,
  method = "nj",
  lasso.pen = "fixed",
  constant = min(sqrt(2 * ncol(X)/(3 * length(y))), 1)
)
}
\arguments{
\item{y}{response vector.}

\item{X}{predictor matrix (do not center or standardize).}

\item{K}{number of groups (minimum and default is two).}

\item{outer}{number of EM starts (default is 10).}

\item{inner}{maximum number of EM iterations (default is 20).}

\item{parallel}{option for running the EM's in parallel, with FALSE being the default. The number of clusters must be specified by the user before calling the \code{rjm} function.}

\item{epsilon}{threshold for termination of the EM, based on the relative change criterion. Default is 1e-06.}

\item{method}{sparse method for the regression component. Available options are 'nj' (default) for the normal-Jeffreys prior and 'lasso'.}

\item{lasso.pen}{option to be specified when \code{method = 'lasso'} for tuning the penalty term. Available options are 'fixed' (default) which uses a two-step cross-validation procedure and 'random' which provides the MAP penalty estimate under a Pareto prior.}

\item{constant}{multiplicative constant (>0) for the lasso penalty when \code{lasso.pen = 'random'}. Default option is min(2p/(3n),1), where p is the number of predictors and n is the sample size.}
}
\value{
An object with S3 class 'regjmix' allowing for call to generic functions such as \code{\link{coef}}.

An object of class 'regjmix' is a list containing the following components:
\item{mu}{matrix of the group-wise means of the predictor variables.}
\item{Sigma}{list of the group-wise covariance matrices of the predictor variables.}
\item{coefficients}{matrix of the group-wise intercepts and regression coefficients.}
\item{lambda}{the group-wise lasso penalties (when \code{method = 'lasso'}).}
\item{sigma2}{the group-wise error variances.}
\item{tau}{the group probabilities.}
\item{z}{the group allocations of the observations.}
\item{model}{summary providing number of groups, the maximum log-likelihood value, number of non-zero parameters and values of AIC and BIC.}
}
\description{
Function for regularized joint mixture (RJM) modeling. For details see
Perrakis, Lartigue, Dondelinger and Mukherjee (2022).
}
\references{
Perrakis, K., Lartigue, T., Dondelinger, F. and  Mukherjee, S. (2022) Regularized joint mixture models. \url{https://arxiv.org/pdf/1908.07869.pdf}
}
\author{
Konstanstinos Perrakis \email{konstantinos.perrakis@durham.ac.uk}
}
